api-function/requirements.txt
azure-functions
pytest